Move `Optimizer` to `wasm-builder`
