+++ gtest/Cargo.toml
@@ -10,7 +10,7 @@ gear-core = { path = "../core" }
 gear-backend-common = { path = "../core-backend/common" }
 gear-backend-wasmtime = { path = "../core-backend/wasmtime" }
 core-processor = { package = "gear-core-processor", path = "../core-processor" }
-wasm-proc = { path = "../utils/wasm-proc" }
+gear-wasm-builder = { path = "../utils/wasm-builder" }
 
 codec = { package = "parity-scale-codec", version = "3.1.2", features = ["derive"] }
 logger = { package = "log", version = "0.4.17" }
