+++ utils/wasm-proc/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 edition = "2018"
 
 [dependencies]
+clap = "3.2"
 log = "0.4.17"
-pwasm-utils = "0.19.0"
+env_logger = "0.9.0"
 gear-wasm-builder = { path = "../wasm-builder" }
