--- utils/wasm-proc/cli/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "wasm-proc-cli"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-wasm-proc = { path = ".." }
-
-clap = "3.2"
-log = "0.4.17"
-env_logger = "0.9.0"
-gear-wasm-builder = { path = "../../wasm-builder" }
-
-[[bin]]
-name = "wasm-proc"
-path = "src/main.rs"
